Decoder

+++ wasm/src/utils.rs
@@ -1,7 +1,7 @@
 use crate::types::Card;
 
 #[allow(dead_code)]
-pub fn make_card(shape: i8, number: i8, color: i8, shade: i8) -> Card {
+pub fn make_card(shape: i32, number: i32, color: i32, shade: i32) -> Card {
     Card {
         shape,
         number,
